Python_App/requirements.txt
bleuio==1.3.1
pyserial==3.5